refactor(parser): split object member parsing helpers
